Add mostly functional front-end
Updated my files to make the front-end UI interact with the game
engine to make the game playable in the browser.

Still to do: Make it so that you can't continue to fill in spaces
after the game ends. Need to transfer player messages from console
to the webpage. Need to highlight the complete row when a player
wins.

Next up: Back end, user accounts, saving stats.