Add API js files and create basic login/logout feature
Created a couple new js files for the API communication, and modified
them to communicate with the API. Basic log in/log out functionality
allows users to track when they are signed in and gives them the option
to sign out or change their password only while logged in.

Next up is to use the API to start tracking games and stats for the
user.