Add basic create game functionality
Updated the api connection to allow games to be created and updated.
Also able to log current games to console, but no user-facing
functionality for that.

Left to do: Add a form that requires player 2 info upon game creation
and pushes that info to the game object. Also need to figure out
a way to list the current games in a user friendly way and allow
users to switch between them.